% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bin_experiment}
\alias{bin_experiment}
\title{Read and join all result files in a directory, appending the names of directories between the dir_experiment and the recorder_dir.}
\usage{
bin_experiment(
  dir_experiment,
  intermediate_dirs,
  thresholds = c(ins_buzz = 0),
  binwidth = 5,
  results_tag = "_buzzdetect",
  append_path = T
)
}
\arguments{
\item{dir_experiment}{The directory holding all buzzdetect results to be analyzed.}

\item{intermediate_dirs}{A vector of strings holding the column names to represent each directory between the experiment directory and the recorder directories. For example, c('site', 'date_deployed') would give the name of the parent directory of the recorder as a 'date_deployed' column, and the name of thegrandparent directory as a 'site' column. description}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron}

\item{binwidth}{The desired width of the bin in minutes}

\item{results_tag}{The tag used to identify result files during file matching, should NOT include a file extension. E.g., _buzzdetect instead of _buzzdetect.csv}
}
\description{
Read and join all result files in a directory, appending the names of directories between the dir_experiment and the recorder_dir.
}
