% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bin_raw}
\alias{bin_raw}
\title{Bin a raw buzzdetect output into bins by time, calling detections according to the input thresholds.}
\usage{
bin_raw(data_raw, thresholds = c(ins_buzz = 0), binwidth = 5, time_start = NA)
}
\arguments{
\item{data_raw}{A data.frame holding raw detection values, in the _buzzdetect.csv format}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron}

\item{binwidth}{The desired width of the bin in minutes}

\item{time_start}{A POSIXct value (or a value that can be converted to such) identifying the date and time that the output begins. Required if the input data has no start_real column, otherwise ignored}
}
\value{
A data.table with detection counts for each neuron listed in thresholds
}
\description{
Bin a raw buzzdetect output into bins by time, calling detections according to the input thresholds.
}
