% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bin_recorder}
\alias{bin_recorder}
\title{Read all files in a recorder directory and return the binned results}
\usage{
bin_recorder(
  dir_recorder,
  intermediate_dirs = NULL,
  thresholds = c(ins_buzz = 0),
  binwidth = 5,
  results_tag = "_buzzdetect"
)
}
\arguments{
\item{dir_recorder}{The file path to the recorder directory}

\item{intermediate_dirs}{A vector of strings holding the column names to represent each directory between the experiment directory and the recorder directories. For example, c('site', 'date_deployed') would give the name of the parent directory of the recorder as a 'date_deployed' column, and the name of thegrandparent directory as a 'site' column. description}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron}

\item{binwidth}{The desired width of the bin in minutes}

\item{results_tag}{The tag used to identify result files during file matching, should NOT include a file extension. E.g., _buzzdetect instead of _buzzdetect.csv}
}
\value{
A data.table
}
\description{
Read all files in a recorder directory and return the binned results
}
