% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{bootstrap_rate}
\alias{bootstrap_rate}
\title{Non-parametric bootstrapping of detection rates}
\usage{
bootstrap_rate(
  dt,
  groupby,
  boot_func = median,
  replicates = 10000,
  conf = 0.95
)
}
\arguments{
\item{dt}{A data.table with a detectionrate column}

\item{groupby}{A vector of column names by which the bootstraps should be grouped}

\item{replicates}{The number of replicates to draw for bootstrapping; passed to the R argument of one.boot}

\item{conf}{The desired confidence interval}
}
\value{
A data.table
}
\description{
Non-parametric bootstrapping of detection rates
}
