% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_log}
\alias{evaluate_log}
\title{Summarize a buzzdetect log file}
\usage{
evaluate_log(path_log)
}
\arguments{
\item{path_log}{String. The path to a buzzdetect results log.}
}
\value{
A named vector of analysis metrics
}
\description{
Evaluate performance in a buzzdetect analysis log. Reports:
\itemize{
\item duration_total - how many seconds did the analysis take?
\item duration_audio - how many seconds of audio were analyzed?
\item analysis_rate - how many seconds of audio were analyzed per second of wall time? NOTE! This does NOT use duration_total as the divisor, it starts counting when the first chunk is reported.
\item bottleneck_time - how long did analyzers wait for streamers to fill the queue?
\item duration_warmup - how long from the launching of analysis until the first chunk was analyzed? Includes results checking and cleaning, streamer launching, etc.
}
}
\details{
Useful for tuning analysis settings.
}
