% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{file_start_time_AudioMoth}
\alias{file_start_time_AudioMoth}
\title{Extract timestamp from path. Assumes YYYYMMDD_HHMMSS format, as used by AudioMoth recorders.}
\usage{
file_start_time_AudioMoth(path_raw, tz = "UTC")
}
\arguments{
\item{path_raw}{Character vector of file paths}

\item{tz}{Timezone string}
}
\value{
POSIXct vector
}
\description{
Extract timestamp from path. Assumes YYYYMMDD_HHMMSS format, as used by AudioMoth recorders.
}
