% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_matching_tag}
\alias{list_matching_tag}
\title{List all files matching a tag (e.g., '_buzzdetect')}
\usage{
list_matching_tag(dir_in, tag)
}
\description{
List all files matching a tag (e.g., '_buzzdetect')
}
