% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording_range.R
\name{plot_recorder_ranges}
\alias{plot_recorder_ranges}
\title{Take input recorder ranges and plot out their runtime}
\usage{
plot_recorder_ranges(ranges)
}
\arguments{
\item{ranges}{A data.frame made by \code{\link{read_file_range}}, or in the same format}
}
\description{
Take input recorder ranges and plot out their runtime
}
