% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{quantile_rate}
\alias{quantile_rate}
\title{Calculate quantiles of detection rates}
\usage{
quantile_rate(dt, groupby, conf = 0.5, median = T)
}
\arguments{
\item{dt}{A data.table with a detectionrate column}

\item{groupby}{A vector of column names by which the quantiles should be grouped}

\item{conf}{The desired confidence interval(s) to calculate quantiles; e.g., 0.95 computes 2.5th and 97.5th percentiles}
}
\value{
A data.table
}
\description{
Calculate quantiles of detection rates
}
