% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{read_dir_range}
\alias{read_dir_range}
\title{@export}
\usage{
read_dir_range(
  path_in,
  parent_dir_names,
  translate_to_real = T,
  tz = NA,
  drop_filetime = T,
  results_tag = TAG_RESULTS
)
}
\description{
@export
}
