% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detections.R
\name{read_raw}
\alias{read_raw}
\title{Read in raw buzzdetect output (this is currently just a wrapper for data.table::fread)}
\usage{
read_raw(path_raw)
}
\arguments{
\item{path_raw}{The file path of the buzzdetect output}
}
\value{
A data.table
}
\description{
Read in raw buzzdetect output (this is currently just a wrapper for data.table::fread)
}
