% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{rebin}
\alias{rebin}
\title{Re-bin already-binned detections.}
\usage{
rebin(data_bin, binwidth)
}
\arguments{
\item{data_bin}{A data.frame in the binned buzzdetect format}

\item{binwidth}{The desired width of the bin in minutes}
}
\description{
Re-bin already-binned detections.
}
