% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{trim}
\alias{trim}
\title{Reduce the size of buzzdetect results by binning at 1 minute, dropping irrelevant columns, and saving as rds
If no thresholds value is supplied, exports neuron activations
When binning into minutes and selecting 2/13 neurons, reduces the size of result files to 1/1,000 the original size.}
\usage{
trim(
  dir_in,
  dir_out,
  thresholds,
  translate_to_real,
  tz = NA,
  delete_original = F,
  overwrite_output = F,
  results_tag = TAG_RESULTS
)
}
\description{
Reduce the size of buzzdetect results by binning at 1 minute, dropping irrelevant columns, and saving as rds
If no thresholds value is supplied, exports neuron activations
When binning into minutes and selecting 2/13 neurons, reduces the size of result files to 1/1,000 the original size.
}
