% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bin}
\alias{bin}
\title{Bin results by time and count detections}
\usage{
bin(results, binwidth, calculate_rate = F)
}
\arguments{
\item{binwidth}{The desired width of the bin in minutes}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron.}

\item{time_start}{Should the detection rate for detection_ columns be calculated? Takes total detections, divides by the number of frames, and outputs as a detectionrate_ column.}
}
\value{
A data.table with a bin_ time column (bin_filetime or bin_realtime), the same detection_ columns as the input, and a frames column counting the total frames in the bin.
}
\description{
Given a results file, group results into bins by time and sum any detection_ columns.
Can re-bin previously binned results, though this gives weird results if the new bin is not a multiple of the old (e.g., bin(10), then bin(15)).
Optionally, calculate the detection rate for detection_ columns (simply the total detections divided by the total frames).
Generally, calculating detection rate is best as a final step. For example, you may first bin into 1 minute bins to compress the dataset,
but then re-bin by 15 minutes for graphing, re-bin by 1 hour for one model and, re-bin by 1 day for another.
}
\details{
Drops all activation columns.
}
