% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bin_directory}
\alias{bin_directory}
\title{Read all buzzdetect results in a directory, apply thresholds to call buzzes, and bin}
\usage{
bin_directory(
  dir_in,
  translate_to_real = T,
  drop_filetime = T,
  parent_dir_names = NULL,
  return_filename = F,
  tz = NA,
  thresholds = c(ins_buzz = 0),
  binwidth = 5,
  results_tag = TAG_RESULTS
)
}
\arguments{
\item{dir_in}{The directory holding all buzzdetect results to be analyzed.}

\item{translate_to_real}{Logical value to choose whether or not to translate start column to real-world times}

\item{parent_dir_names}{A vector of strings holding the column names to associate with the parent directories of the file in question. Unnamed directories will not be returned.}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron}

\item{binwidth}{The desired width of the bin in minutes}

\item{results_tag}{The tag used to identify result files during file matching, should NOT include a file extension. E.g., _buzzdetect instead of _buzzdetect.csv}
}
\value{
A data.table
}
\description{
Read all buzzdetect results in a directory, apply thresholds to call buzzes, and bin
}
