% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{bootstrap_rate}
\alias{bootstrap_rate}
\title{Non-parametric bootstrapping of detection rates}
\usage{
bootstrap_rate(dt, groupby)
}
\arguments{
\item{dt}{A data.table with a detectionrate column}

\item{groupby}{A vector of column names by which the bootstraps should be grouped}
}
\value{
A data.table
}
\description{
Non-parametric bootstrapping of detection rates
}
