% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{call_detections}
\alias{call_detections}
\title{Call detections using thresholds}
\usage{
call_detections(results, thresholds)
}
\arguments{
\item{results}{A data frame of buzzdetect results, such as created by \link{read_results}..
Must have activation_ columns corresponding to the neurons named in thresholds
and must \emph{not} have detections_ column for the same.}

\item{thresholds}{A named numeric vector with names corresponding to neuron names and values corresponding to the desired detection threshold for that neuron.}
}
\description{
Call detections using thresholds
}
\examples{
call_detections(
  results = buzzr::read_results('models/model_general_v3/output/testbuzz_buzzdetect.csv'),
  thresholds = c(ins_buzz = -1.2, ambient_rain = -1)
)
}
