% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{commontime}
\alias{commontime}
\title{Set the year, month, and date of input datetimes to 2000-01-01. Used for plotting results from different time periods on a common X axis.}
\usage{
commontime(times, tz = Sys.timezone())
}
\arguments{
\item{times}{Vector of POSIXct datetimes}
}
\description{
Set the year, month, and date of input datetimes to 2000-01-01. Used for plotting results from different time periods on a common X axis.
}
