% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{file_start_time}
\alias{file_start_time}
\title{Extract timestamp from path.}
\usage{
file_start_time(paths, tz = "America/New_York")
}
\arguments{
\item{paths}{Character vector of file paths}

\item{tz}{Timezone string}
}
\value{
POSIXct vector
}
\description{
Extract timestamp from path.
}
