% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_start_time}
\alias{file_start_time}
\title{Given a path, extract the start time as POSIXct object. Assumes YYMMDD_HHMM format, as used by Sony ICD-PX370 recorders.}
\usage{
file_start_time(path_raw)
}
\arguments{
\item{path_raw}{The file path of any file with a name starting with a YYMMDD_HHMM timestamp}
}
\value{
A POSIXct object bearing the start time of the file
}
\description{
Given a path, extract the start time as POSIXct object. Assumes YYMMDD_HHMM format, as used by Sony ICD-PX370 recorders.
}
