% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{file_start_time}
\alias{file_start_time}
\title{Extract timestamp from path.}
\usage{
file_start_time(paths, tz)
}
\arguments{
\item{paths}{Character vector of file paths}

\item{tz}{What time zone should the timestamp be interpreted as? See \link[base:timezones]{base::timezones}}
}
\value{
POSIXct vector
}
\description{
Extract timestamp from path.
}
