% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_start_time}
\alias{file_start_time}
\title{Extract timestamp from path. Assumes YYMMDD_HHMM format, as used by Sony ICD-PX370 recorders.}
\usage{
file_start_time(path_raw, tz = "America/New_York")
}
\arguments{
\item{path_raw}{Character vector of file paths}

\item{tz}{Timezone string}
}
\value{
POSIXct vector
}
\description{
Extract timestamp from path. Assumes YYMMDD_HHMM format, as used by Sony ICD-PX370 recorders.
}
