% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{palette}
\alias{palette}
\title{The buzzdetect color palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette
}
\description{
This is named, discrete version of a magma color palette (magma is frequently used for spectrograms).
See \link[viridis:reexports]{viridis::magma} for a continuous version.
}
\keyword{datasets}
