% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_directory}
\alias{read_directory}
\title{Read all buzzdetect results in a directory, recursively}
\usage{
read_directory(
  dir_results,
  translate_to_real = T,
  drop_filetime = T,
  parent_dir_names = NULL,
  return_ident = F,
  tz = NA,
  results_tag = "_buzzdetect"
)
}
\arguments{
\item{dir_results}{The directory holding all buzzdetect results to be read.}

\item{translate_to_real}{Logical value to choose whether or not to translate start column to real-world times.}

\item{parent_dir_names}{A vector of strings holding the column names to associate with the parent directories of the file in question. Unnamed directories will not be returned.}

\item{return_ident}{Whether or not to include an ident column. The ident is the relative path from your results directory to the file, not including the file extension or results tag. For example, if your results directory is models/model_general_v3/output and the result file is './foo/bar/testbuzz_buzzdetect.csv', the ident is 'foo/bar/testbuzz'.}

\item{tz}{The time zone for the results, as in \link[base:as.POSIXlt]{base::as.POSIXct}.}

\item{results_tag}{The tag used to identify result files during file matching, should NOT include a file extension. E.g., '_buzzdetect' instead of '_buzzdetect.csv'.}

\item{return_filename}{A boolean representing whether or not to return the filename as an element/column.}
}
\description{
Applies \link{read_results} to an entire directory, joining results.
optionally adding columns for parent directories.
}
