% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_directory}
\alias{read_directory}
\title{Read and join all result files in a directory, optionally adding columns for parent directories.}
\usage{
read_directory(
  dir_in,
  translate_to_real = T,
  drop_filetime = T,
  parent_dir_names = NULL,
  return_filename = F,
  tz = "UTC"
)
}
\arguments{
\item{dir_in}{The directory holding all buzzdetect results to be analyzed.}

\item{translate_to_real}{Logical value to choose whether or not to translate start column to real-world times}

\item{parent_dir_names}{A vector of strings holding the column names to associate with the parent directories of the file in question. Unnamed directories will not be returned.}

\item{results_tag}{The tag used to identify result files during file matching, should NOT include a file extension. E.g., _buzzdetect instead of _buzzdetect.csv}
}
\description{
Read and join all result files in a directory, optionally adding columns for parent directories.
}
