% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{read_file_range}
\alias{read_file_range}
\title{Get the start and end times for a raw result}
\usage{
read_file_range(path_in, translate_to_real = T, tz = NA, drop_filetime = T)
}
\value{
A data.frame with the following columns:
recorder: recorder IDs
timepoint: either "start" or "end", denoting whether it's the first or last frame from the recorder
time: a date-time corresponding to the start or end time
file: the filename or unique identifier for each file from a recorder
@export
}
\description{
Get the start and end times for a raw result
}
