% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_raw}
\alias{read_raw}
\title{Read in raw buzzdetect output and optionally translate start column to real-world times.}
\usage{
read_raw(path_raw, translate_to_real = T, tz = NA, drop_filetime = T)
}
\arguments{
\item{path_raw}{The file path to the raw buzzdetect output file. The filetype can be csv or rds.}

\item{translate_to_real}{Logical value to choose whether or not to translate start column to real-world times}
}
\value{
A data.table of the raw data file; the start column will be replaced with start_real if translate_to_real is set to TRUE.
}
\description{
Read in raw buzzdetect output and optionally translate start column to real-world times.
}
