% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results}
\alias{read_results}
\title{Read in raw result files}
\usage{
read_results(path_in, translate_to_real = T, tz = NA, drop_filetime = T)
}
\arguments{
\item{translate_to_real}{Logical value to choose whether or not to translate start column to real-world times}

\item{path_raw}{The file path to the raw buzzdetect output file. The filetype can be csv or rds.}
}
\description{
Read in raw result files
}
