% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detections.R
\name{rebin}
\alias{rebin}
\title{Re-bin existing binned detections}
\usage{
rebin(dt_bin, binwidth)
}
\arguments{
\item{dt_bin}{A data.table in the binned buzzdetect format}

\item{binwidth}{The desired with of the bin in minutes, to be passed to the "unit" argument of lubridate::floor_date()}
}
\value{
A data.table
}
\description{
Re-bin existing binned detections
}
