% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_of_day}
\alias{time_of_day}
\title{Convert a date-time to a time of day as proportion (0,1) or hour (0,24; with decimals).}
\usage{
time_of_day(times, time_format = "proportion")
}
\arguments{
\item{times}{A vector of POSIX times.}
}
\description{
Useful for statistical analyses, e.g. modeling the time of peak foraging.
For plotting, use \link{commontime} and \link{label_hour}.
}
