% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_of_day}
\alias{time_of_day}
\title{Get the time of day as a 0-1 value (left-inclusive) where 0 represents midnight}
\usage{
time_of_day(times, time_format = "proportion")
}
\arguments{
\item{times}{A vector of POSIX times.}
}
\description{
Get the time of day as a 0-1 value (left-inclusive) where 0 represents midnight
}
